refactor layout tree based
