+++ FSCruiserV2/FSCruiserV2CE_CF20.csproj
@@ -223,6 +223,12 @@
     <Compile Include="Core\Models\TallySettingsDO.cs" />
     <Compile Include="Core\Models\TreeVM.cs" />
     <Compile Include="NetCF\WinForms\ViewController.cs" />
+    <Compile Include="WinForms.Common\TreeBasedTallyView_Base.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="WinForms.Common\TreeBasedTallyView_Base.Designer.cs">
+      <DependentUpon>TreeBasedTallyView_Base.cs</DependentUpon>
+    </Compile>
     <Compile Include="WinForms.Common\ViewController_Base.cs" />
     <Compile Include="NetCF\Program.cs" />
     <Compile Include="Properties\AssemblyInfoCE_CF20.cs" />
@@ -280,10 +286,10 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <Compile Include="NetCF\WinForms\TallyRow.cs">
+    <Compile Include="NetCF\WinForms\DataEntry\TallyRow.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="NetCF\WinForms\TallyRow.Designer.cs">
+    <Compile Include="NetCF\WinForms\DataEntry\TallyRow.Designer.cs">
       <DependentUpon>TallyRow.cs</DependentUpon>
     </Compile>
     <Compile Include="NetCF\Win32.cs" />
@@ -334,9 +340,12 @@
       <DependentUpon>FormTallySetup.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
-    <EmbeddedResource Include="NetCF\WinForms\TallyRow.resx">
+    <EmbeddedResource Include="NetCF\WinForms\DataEntry\TallyRow.resx">
       <DependentUpon>TallyRow.cs</DependentUpon>
       <SubType>Designer</SubType>
+    </EmbeddedResource>
+    <EmbeddedResource Include="WinForms.Common\TreeBasedTallyView_Base.resx">
+      <DependentUpon>TreeBasedTallyView_Base.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
